requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudinary==1.34.0
colorama==0.4.6
cryptography==41.0.4
cssselect2==0.7.0
Django==4.2.5
django-cloudinary-storage==0.3.0
django-environ==0.11.2
djangorestframework==3.14.0
html5lib==1.1
idna==3.4
lxml==4.9.3
oscrypto==1.3.0
Pillow==10.0.0
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pypdf==3.16.1
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.2
urllib3==1.26.16
webencodings==0.5.1
whitenoise==6.5.0
xhtml2pdf==0.2.11
